post service code refactoring and separate comment service devide and shorten code as well
