+++ src/main/java/com/example/demo/repository/implementation/PostRepositoryImpl.java
@@ -43,14 +43,6 @@ public class PostRepositoryImpl implements PostRepository {
     private final LikedRecipePostRowMapper likedRecipePostRowMapper;
     private final LikedIMadePostRowMapper likedIMadePostRowMapper;
     private final LikedOtherPostRowMapper likedOtherPostRowMapper;
-    private final LikedGardeningCommRowMapper likedGardeningCommRowMapper;
-    private final LikedRecipeCommRowMapper likedRecipeCommRowMapper;
-    private final LikedIMadeCommRowMapper likedIMadeCommRowMapper;
-    private final LikedOtherCommRowMapper likedOtherCommRowMapper;
-    private final GardeningCommentRowMapper gardeningCommentRowMapper;
-    private final RecipeCommentRowMapper recipeCommentRowMapper;
-    private final IMadeCommentRowMapper iMadeCommentRowMapper;
-    private final OtherCommentRowMapper otherCommentRowMapper;
     private final RecipePostRowMapper recipePostRowMapper;
     private final IMadePostRowMapper iMadePostRowMapper;
     private final OtherPostRowMapper otherPostRowMapper;
@@ -138,172 +130,40 @@ public class PostRepositoryImpl implements PostRepository {
     }
 
     @Override
-    public GardeningComment addGardeningComment(Long userId, Long postId, GardeningComment gardeningComment) {
+    public void addPostLikeKeyTable(Long userId, Long postId, PostType postType) {
         try {
-            LocalDateTime currentDateTime = LocalDateTime.now();
-            MapSqlParameterSource parameters = new MapSqlParameterSource();
-            parameters.addValue("date", Timestamp.valueOf(currentDateTime));
-            parameters.addValue("comment_text", gardeningComment.getComment_text());
-            parameters.addValue("reply_username", gardeningComment.getReply_username());
-            parameters.addValue("likes", 0);
-            parameters.addValue("parent_comment_id", gardeningComment.getParent_comment_id());
-            parameters.addValue("comment_user_id",userId);
-            parameters.addValue("comment_gardening_post_id", postId);
-            KeyHolder keyHolder = new GeneratedKeyHolder();
-            jdbc.update(INSERT_GARDENING_COMMENT_QUERY, parameters, keyHolder);
-            long generatedId = Objects.requireNonNull(keyHolder.getKey()).longValue();
-            gardeningComment.setId(generatedId);
-            gardeningComment.setDate(currentDateTime);
-            return gardeningComment;
-        } catch (Exception exception){
-            throw new ApiException("An error occurred. Please try again");
-        }
-    }
-
-    @Override
-    public RecipeComment addRecipeComment(Long userId, Long postId, RecipeComment recipeComment) {
-        try {
-            LocalDateTime currentDateTime = LocalDateTime.now();
-            MapSqlParameterSource parameters = new MapSqlParameterSource();
-            parameters.addValue("date", Timestamp.valueOf(currentDateTime));
-            parameters.addValue("comment_text", recipeComment.getComment_text());
-            parameters.addValue("reply_username", recipeComment.getReply_username());
-            parameters.addValue("likes", 0);
-            parameters.addValue("parent_comment_id", recipeComment.getParent_comment_id());
-            parameters.addValue("comment_user_id",userId);
-            parameters.addValue("comment_recipe_post_id", postId);
-            KeyHolder keyHolder = new GeneratedKeyHolder();
-            jdbc.update(INSERT_RECIPE_COMMENT_QUERY, parameters, keyHolder);
-            long generatedId = Objects.requireNonNull(keyHolder.getKey()).longValue();
-            recipeComment.setId(generatedId);
-            recipeComment.setDate(currentDateTime);
-            return recipeComment;
-        } catch (Exception exception){
-            throw new ApiException("An error occurred. Please try again");
-        }
-    }
-    @Override
-    public IMadeComment addIMadeComment(Long userId, Long postId, IMadeComment iMadeComment) {
-        try {
-            LocalDateTime currentDateTime = LocalDateTime.now();
-            MapSqlParameterSource parameters = new MapSqlParameterSource();
-            parameters.addValue("date", Timestamp.valueOf(currentDateTime));
-            parameters.addValue("comment_text", iMadeComment.getComment_text());
-            parameters.addValue("reply_username", iMadeComment.getReply_username());
-            parameters.addValue("likes", 0);
-            parameters.addValue("parent_comment_id", iMadeComment.getParent_comment_id());
-            parameters.addValue("comment_user_id",userId);
-            parameters.addValue("comment_i_made_post_id", postId);
-            KeyHolder keyHolder = new GeneratedKeyHolder();
-            jdbc.update(INSERT_I_MADE_COMMENT_QUERY, parameters, keyHolder);
-            long generatedId = Objects.requireNonNull(keyHolder.getKey()).longValue();
-            iMadeComment.setId(generatedId);
-            iMadeComment.setDate(currentDateTime);
-            return iMadeComment;
-        } catch (Exception exception){
-            throw new ApiException("An error occurred. Please try again");
-        }
-    }
-    @Override
-    public OtherComment addOtherComment(Long userId, Long postId, OtherComment otherComment) {
-        try {
-            LocalDateTime currentDateTime = LocalDateTime.now();
-            MapSqlParameterSource parameters = new MapSqlParameterSource();
-            parameters.addValue("date", Timestamp.valueOf(currentDateTime));
-            parameters.addValue("comment_text", otherComment.getComment_text());
-            parameters.addValue("reply_username", otherComment.getReply_username());
-            parameters.addValue("likes", 0);
-            parameters.addValue("parent_comment_id", otherComment.getParent_comment_id());
-            parameters.addValue("comment_user_id",userId);
-            parameters.addValue("comment_other_post_id", postId);
-            KeyHolder keyHolder = new GeneratedKeyHolder();
-            jdbc.update(INSERT_OTHER_COMMENT_QUERY, parameters, keyHolder);
-            long generatedId = Objects.requireNonNull(keyHolder.getKey()).longValue();
-            otherComment.setId(generatedId);
-            otherComment.setDate(currentDateTime);
-            return otherComment;
-        } catch (Exception exception){
-            throw new ApiException("An error occurred. Please try again");
-        }
-    }
-
-    @Override
-    public List<GardeningComment> getAllGardeningCommentsByPostId(Long postId) {
-        try {
-            Map<String, Object> params = new HashMap<>();
-            params.put("comment_gardening_post_id", postId);
-            return jdbc.query(SELECT_ALL_GARDENING_COMMENTS_BY_POST_ID, params, gardeningCommentRowMapper);
+            String addQuery = getAddPostLikeQuery(postType);
+            jdbc.update(addQuery, of("userId", userId, "postId", postId));
         } catch (Exception exception) {
-            log.error(exception.getMessage());
-            throw new ApiException("An error occurred. Please try again");
-        }
-    }
-
-    @Override
-    public List<RecipeComment> getAllRecipeCommentsByPostId(Long postId) {
-        try {
-            Map<String, Object> params = new HashMap<>();
-            params.put("comment_recipe_post_id", postId);
-            return jdbc.query(SELECT_ALL_RECIPE_COMMENTS_BY_POST_ID, params, recipeCommentRowMapper);
-        } catch (Exception exception) {
-            log.error(exception.getMessage());
-            throw new ApiException("An error occurred. Please try again");
-        }
-    }
-
-    @Override
-    public List<IMadeComment> getAllIMadeCommentsByPostId(Long postId) {
-        try {
-            Map<String, Object> params = new HashMap<>();
-            params.put("comment_i_made_post_id", postId);
-            return jdbc.query(SELECT_ALL_I_MADE_COMMENTS_BY_POST_ID, params, iMadeCommentRowMapper);
-        } catch (Exception exception) {
-            log.error(exception.getMessage());
-            throw new ApiException("An error occurred. Please try again");
-        }
-    }
-
-    @Override
-    public List<OtherComment> getAllOtherCommentsByPostId(Long postId) {
-        try {
-            Map<String, Object> params = new HashMap<>();
-            params.put("comment_other_post_id", postId);
-            return jdbc.query(SELECT_ALL_OTHER_COMMENTS_BY_POST_ID, params, otherCommentRowMapper);
-        } catch (Exception exception) {
-            log.error(exception.getMessage());
-            throw new ApiException("An error occurred. Please try again");
-        }
-    }
-
-    @Override
-    public void updateComment(Long commentId, String comment_text, PostType postType) {
-        try {
-        switch (postType) {
-            case GARDENING -> jdbc.update(UPDATE_GARDENING_COMMENT_BY_COMMENT_ID_QUERY, of("id", commentId, "comment_text", comment_text));
-            case RECIPE -> jdbc.update(UPDATE_RECIPE_COMMENT_BY_COMMENT_ID_QUERY, of("id", commentId, "comment_text", comment_text));
-            case I_MADE -> jdbc.update(UPDATE_I_MADE_COMMENT_BY_COMMENT_ID_QUERY, of("id", commentId, "comment_text", comment_text));
-            case OTHER -> jdbc.update(UPDATE_OTHER_COMMENT_BY_COMMENT_ID_QUERY, of("id", commentId, "comment_text", comment_text));
-            default -> throw new IllegalArgumentException("Unsupported post type: " + postType);
-        }
-            } catch (Exception exception) {
-                throw new ApiException("An error occurred. Please try again.");
-            }
-    }
-    @Override
-    public GardeningComment getLatestGardeningComment(Long postId) {
-        try {
-            return jdbc.queryForObject(SELECT_LATEST_COMMENT_FROM_GARDENING_POST, of("postId", postId), gardeningCommentRowMapper);
-        } catch (EmptyResultDataAccessException exception) {
-            throw new ApiException("Comments were not fount: " + postId);
-        } catch (Exception exception) {
-            log.error(exception.getMessage());
             throw new ApiException("An error occurred. Please try again.");
         }
     }
 
     @Override
-    public void deleteComment(Long commentId, PostType postType) {
-        jdbcUpdateQueryWithIdParameter(commentId, postType, DELETE_GARDENING_COMMENT, DELETE_RECIPE_COMMENT, DELETE_I_MADE_COMMENT, DELETE_OTHER_COMMENT);
+    public void deletePostLikeKeyTable(Long userId, Long postId, PostType postType) {
+        try {
+            String deleteQuery = getDeletePostLikeQuery(postType);
+            jdbc.update(deleteQuery, of("userId", userId, "postId", postId));
+        } catch (Exception exception) {
+            throw new ApiException("An error occurred. Please try again.");
+        }
+    }
+
+    private String getAddPostLikeQuery(PostType postType) {
+        return getPostString(postType, ADD_GARDENING_POST_LIKES_KEY_TABLE, ADD_RECIPE_POST_LIKES_KEY_TABLE, ADD_I_MADE_POST_LIKES_KEY_TABLE, ADD_OTHER_POST_LIKES_KEY_TABLE);
+    }
+    private String getDeletePostLikeQuery(PostType postType) {
+        return getPostString(postType, DELETE_GARDENING_POST_LIKES_KEY_TABLE, DELETE_RECIPE_POST_LIKES_KEY_TABLE, DELETE_I_MADE_POST_LIKES_KEY_TABLE, DELETE_OTHER_POST_LIKES_KEY_TABLE);
+    }
+
+    private String getPostString(PostType postType, String GardeningTable, String RecipeTable, String IMadeTable, String OtherTable) {
+        return switch (postType) {
+            case GARDENING -> GardeningTable;
+            case RECIPE -> RecipeTable;
+            case I_MADE -> IMadeTable;
+            case OTHER -> OtherTable;
+            default -> throw new IllegalArgumentException("Unsupported comment type: " + postType);
+        };
     }
 
     @Override
@@ -314,16 +174,6 @@ public class PostRepositoryImpl implements PostRepository {
     @Override
     public void removePostLike(Long id, PostType postType) {
             jdbcUpdateQueryWithIdParameter(id, postType, UPDATE_MINUS_GARDENING_LIKES, UPDATE_MINUS_RECIPE_LIKES, UPDATE_MINUS_I_MADE_LIKES, UPDATE_MINUS_OTHER_LIKES);
-    }
-
-    @Override
-    public void addCommentLike(Long id, PostType postType) {
-        jdbcUpdateQueryWithIdParameter(id, postType, UPDATE_PLUS_GARDENING_COMMENT_LIKES, UPDATE_PLUS_RECIPE_COMMENT_LIKES, UPDATE_PLUS_I_MADE_COMMENT_LIKES, UPDATE_PLUS_OTHER_COMMENT_LIKES);
-    }
-
-    @Override
-    public void removeCommentLike(Long id, PostType postType) {
-        jdbcUpdateQueryWithIdParameter(id, postType, UPDATE_MINUS_GARDENING_COMMENT_LIKES, UPDATE_MINUS_RECIPE_COMMENT_LIKES, UPDATE_MINUS_I_MADE_COMMENT_LIKES, UPDATE_MINUS_OTHER_COMMENT_LIKES);
     }
 
     private void jdbcUpdateQueryWithIdParameter(Long id, PostType postType, String updateGardening, String updateRecipe, String updateIMade, String updateOther) {
@@ -341,27 +191,9 @@ public class PostRepositoryImpl implements PostRepository {
     }
 
     @Override
-    public void addGardeningPostLikeKeyTable(Long userId, Long postId) {
+    public boolean userHasLikedPost(Long userId, Long postId, String tableName) {
         try {
-            jdbc.update(ADD_GARDENING_POST_LIKES_KEY_TABLE, of("userId", userId, "postId", postId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred. Please try again.");
-        }
-    }
-
-    @Override
-    public void deleteGardeningPostLikeKeyTable(Long userId, Long postId) {
-        try {
-            jdbc.update(DELETE_GARDENING_POST_LIKES_KEY_TABLE, of("userId", userId, "postId", postId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred. Please try again.");
-        }
-    }
-
-    @Override
-    public boolean userHasLikedGardeningPost(Long userId, Long postId) {
-        try {
-            String sql = "SELECT COUNT(*) FROM GardeningPostLikes WHERE user_id = :userId AND post_id = :postId";
+            String sql = String.format("SELECT COUNT(*) FROM %s WHERE user_id = :userId AND post_id = :postId", tableName);
             SqlParameterSource params = new MapSqlParameterSource()
                     .addValue("userId", userId)
                     .addValue("postId", postId);
@@ -371,7 +203,6 @@ public class PostRepositoryImpl implements PostRepository {
             throw new ApiException("An error occurred while checking if the user has liked the post.");
         }
     }
-
     @Override
     public List<LikedGardeningPost> getUserLikedGardeningPosts(Long userId) {
         try {
@@ -387,86 +218,6 @@ public class PostRepositoryImpl implements PostRepository {
             throw new ApiException("An error occurred while retrieving liked posts.");
         }
     }
-
-    @Override
-    public void addGardeningCommentLikeKeyTable(Long userId, Long commentId) {
-        try {
-            jdbc.update(ADD_GARDENING_COMMENT_LIKES_KEY_TABLE, of("userId", userId, "commentId", commentId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred. Please try again.");
-        }
-    }
-
-    @Override
-    public void deleteGardeningCommentLikeKeyTable(Long userId, Long commentId) {
-        try {
-            jdbc.update(DELETE_GARDENING_COMMENT_LIKES_KEY_TABLE, of("userId", userId, "commentId", commentId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred. Please try again.");
-        }
-    }
-
-    @Override
-    public boolean userHasLikedGardeningComment(Long userId, Long commentId) {
-        try {
-            String sql = "SELECT COUNT(*) FROM GardeningCommentLikes WHERE user_id = :userId AND comment_id = :commentId";
-            SqlParameterSource params = new MapSqlParameterSource()
-                    .addValue("userId", userId)
-                    .addValue("commentId", commentId);
-            int likeCount = jdbc.queryForObject(sql, params, Integer.class);
-            return likeCount > 0;
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred while checking if the user has liked the comment.");
-        }
-    }
-
-    @Override
-    public List<LikedGardeningComment> getUserLikedGardeningComments(Long userId) {
-        try {
-            String sql = "SELECT gc.id, gc.likes " +
-                    "FROM GardeningCommentLikes cl " +
-                    "INNER JOIN GardeningComment gc ON cl.comment_id = gc.id " +
-                    "WHERE cl.user_id = :userId";
-            List<LikedGardeningComment> likedComments = jdbc.query(sql, of("userId", userId), likedGardeningCommRowMapper);
-            return likedComments;
-        } catch (Exception exception) {
-            log.error("Error retrieving liked comments for user " + userId, exception);
-            throw new ApiException("An error occurred while retrieving liked comments.");
-        }
-    }
-
-    @Override
-    public void addRecipePostLikeKeyTable(Long userId, Long postId) {
-        try {
-            jdbc.update(ADD_RECIPE_POST_LIKES_KEY_TABLE, of("userId", userId, "postId", postId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred. Please try again.");
-        }
-    }
-
-    @Override
-    public void deleteRecipePostLikeKeyTable(Long userId, Long postId) {
-        try {
-            jdbc.update(DELETE_RECIPE_POST_LIKES_KEY_TABLE, of("userId", userId, "postId", postId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred. Please try again.");
-        }
-    }
-
-    @Override
-    public boolean userHasLikedRecipePost(Long userId, Long postId) {
-        try {
-            String sql = "SELECT COUNT(*) FROM RecipePostLikes WHERE user_id = :userId AND post_id = :postId";
-            SqlParameterSource params = new MapSqlParameterSource()
-                    .addValue("userId", userId)
-                    .addValue("postId", postId);
-            int likeCount = jdbc.queryForObject(sql, params, Integer.class);
-            return likeCount > 0;
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred while checking if the user has liked the post.");
-        }
-    }
-
     @Override
     public List<LikedRecipePost> getUserLikedRecipePosts(Long userId) {
         try {
@@ -483,85 +234,6 @@ public class PostRepositoryImpl implements PostRepository {
         }
     }
     @Override
-    public void addRecipeCommentLikeKeyTable(Long userId, Long commentId) {
-        try {
-            jdbc.update(ADD_RECIPE_COMMENT_LIKES_KEY_TABLE, of("userId", userId, "commentId", commentId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred maybe. Please try again.");
-        }
-    }
-
-    @Override
-    public void deleteRecipeCommentLikeKeyTable(Long userId, Long commentId) {
-        try {
-            jdbc.update(DELETE_RECIPE_COMMENT_LIKES_KEY_TABLE, of("userId", userId, "commentId", commentId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred some. Please try again.");
-        }
-    }
-
-    @Override
-    public boolean userHasLikedRecipeComment(Long userId, Long commentId) {
-        try {
-            String sql = "SELECT COUNT(*) FROM RecipeCommentLikes WHERE user_id = :userId AND comment_id = :commentId";
-            SqlParameterSource params = new MapSqlParameterSource()
-                    .addValue("userId", userId)
-                    .addValue("commentId", commentId);
-            int likeCount = jdbc.queryForObject(sql, params, Integer.class);
-            return likeCount > 0;
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred while checking if the user has liked the comment.");
-        }
-    }
-
-    @Override
-    public List<LikedRecipeComment> getUserLikedRecipeComments(Long userId) {
-        try {
-            String sql = "SELECT gc.id, gc.likes " +
-                    "FROM RecipeCommentLikes cl " +
-                    "INNER JOIN RecipeComment gc ON cl.comment_id = gc.id " +
-                    "WHERE cl.user_id = :userId";
-            List<LikedRecipeComment> likedComments = jdbc.query(sql, of("userId", userId), likedRecipeCommRowMapper);
-            return likedComments;
-        } catch (Exception exception) {
-            log.error("Error retrieving liked comments for user " + userId, exception);
-            throw new ApiException("An error occurred while retrieving liked comments.");
-        }
-    }
-
-    @Override
-    public void addIMadePostLikeKeyTable(Long userId, Long postId) {
-        try {
-            jdbc.update(ADD_I_MADE_POST_LIKES_KEY_TABLE, of("userId", userId, "postId", postId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred. Please try again.");
-        }
-    }
-
-    @Override
-    public void deleteIMadePostLikeKeyTable(Long userId, Long postId) {
-        try {
-            jdbc.update(DELETE_I_MADE_POST_LIKES_KEY_TABLE, of("userId", userId, "postId", postId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred. Please try again.");
-        }
-    }
-
-    @Override
-    public boolean userHasLikedIMadePost(Long userId, Long postId) {
-        try {
-            String sql = "SELECT COUNT(*) FROM IMadePostLikes WHERE user_id = :userId AND post_id = :postId";
-            SqlParameterSource params = new MapSqlParameterSource()
-                    .addValue("userId", userId)
-                    .addValue("postId", postId);
-            int likeCount = jdbc.queryForObject(sql, params, Integer.class);
-            return likeCount > 0;
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred while checking if the user has liked the post.");
-        }
-    }
-
-    @Override
     public List<LikedIMadePost> getUserLikedIMadePosts(Long userId) {
         try {
             String sql = "SELECT gp.id, gp.likes " +
@@ -576,86 +248,6 @@ public class PostRepositoryImpl implements PostRepository {
             throw new ApiException("An error occurred while retrieving liked posts.");
         }
     }
-
-    @Override
-    public void addIMadeCommentLikeKeyTable(Long userId, Long commentId) {
-        try {
-            jdbc.update(ADD_I_MADE_COMMENT_LIKES_KEY_TABLE, of("userId", userId, "commentId", commentId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred. Please try again.");
-        }
-    }
-
-    @Override
-    public void deleteIMadeCommentLikeKeyTable(Long userId, Long commentId) {
-        try {
-            jdbc.update(DELETE_I_MADE_COMMENT_LIKES_KEY_TABLE, of("userId", userId, "commentId", commentId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred. Please try again.");
-        }
-    }
-
-    @Override
-    public boolean userHasLikedIMadeComment(Long userId, Long commentId) {
-        try {
-            String sql = "SELECT COUNT(*) FROM IMadeCommentLikes WHERE user_id = :userId AND comment_id = :commentId";
-            SqlParameterSource params = new MapSqlParameterSource()
-                    .addValue("userId", userId)
-                    .addValue("commentId", commentId);
-            int likeCount = jdbc.queryForObject(sql, params, Integer.class);
-            return likeCount > 0;
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred while checking if the user has liked the comment.");
-        }
-    }
-
-    @Override
-    public List<LikedIMadeComment> getUserLikedIMadeComments(Long userId) {
-        try {
-            String sql = "SELECT gc.id, gc.likes " +
-                    "FROM IMadeCommentLikes cl " +
-                    "INNER JOIN IMadeComment gc ON cl.comment_id = gc.id " +
-                    "WHERE cl.user_id = :userId";
-            List<LikedIMadeComment> likedComments = jdbc.query(sql, of("userId", userId), likedIMadeCommRowMapper);
-            return likedComments;
-        } catch (Exception exception) {
-            log.error("Error retrieving liked comments for user " + userId, exception);
-            throw new ApiException("An error occurred while retrieving liked comments.");
-        }
-    }
-
-    @Override
-    public void addOtherPostLikeKeyTable(Long userId, Long postId) {
-        try {
-            jdbc.update(ADD_OTHER_POST_LIKES_KEY_TABLE, of("userId", userId, "postId", postId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred. Please try again.");
-        }
-    }
-
-    @Override
-    public void deleteOtherPostLikeKeyTable(Long userId, Long postId) {
-        try {
-            jdbc.update(DELETE_OTHER_POST_LIKES_KEY_TABLE, of("userId", userId, "postId", postId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred. Please try again.");
-        }
-    }
-
-    @Override
-    public boolean userHasLikedOtherPost(Long userId, Long postId) {
-        try {
-            String sql = "SELECT COUNT(*) FROM OtherPostLikes WHERE user_id = :userId AND post_id = :postId";
-            SqlParameterSource params = new MapSqlParameterSource()
-                    .addValue("userId", userId)
-                    .addValue("postId", postId);
-            int likeCount = jdbc.queryForObject(sql, params, Integer.class);
-            return likeCount > 0;
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred while checking if the user has liked the post.");
-        }
-    }
-
     @Override
     public List<LikedOtherPost> getUserLikedOtherPosts(Long userId) {
         try {
@@ -669,53 +261,6 @@ public class PostRepositoryImpl implements PostRepository {
         } catch (Exception exception) {
             log.error("Error retrieving liked posts for user " + userId, exception);
             throw new ApiException("An error occurred while retrieving liked posts.");
-        }
-    }
-
-    @Override
-    public void addOtherCommentLikeKeyTable(Long userId, Long commentId) {
-        try {
-            jdbc.update(ADD_OTHER_COMMENT_LIKES_KEY_TABLE, of("userId", userId, "commentId", commentId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred. Please try again.");
-        }
-    }
-
-    @Override
-    public void deleteOtherCommentLikeKeyTable(Long userId, Long commentId) {
-        try {
-            jdbc.update(DELETE_OTHER_COMMENT_LIKES_KEY_TABLE, of("userId", userId, "commentId", commentId));
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred. Please try again.");
-        }
-    }
-
-    @Override
-    public boolean userHasLikedOtherComment(Long userId, Long commentId) {
-        try {
-            String sql = "SELECT COUNT(*) FROM OtherCommentLikes WHERE user_id = :userId AND comment_id = :commentId";
-            SqlParameterSource params = new MapSqlParameterSource()
-                    .addValue("userId", userId)
-                    .addValue("commentId", commentId);
-            int likeCount = jdbc.queryForObject(sql, params, Integer.class);
-            return likeCount > 0;
-        } catch (Exception exception) {
-            throw new ApiException("An error occurred while checking if the user has liked the comment.");
-        }
-    }
-
-    @Override
-    public List<LikedOtherComment> getUserLikedOtherComments(Long userId) {
-        try {
-            String sql = "SELECT gc.id, gc.likes " +
-                    "FROM OtherCommentLikes cl " +
-                    "INNER JOIN OtherComment gc ON cl.comment_id = gc.id " +
-                    "WHERE cl.user_id = :userId";
-            List<LikedOtherComment> likedComments = jdbc.query(sql, of("userId", userId), likedOtherCommRowMapper);
-            return likedComments;
-        } catch (Exception exception) {
-            log.error("Error retrieving liked comments for user " + userId, exception);
-            throw new ApiException("An error occurred while retrieving liked comments.");
         }
     }
 
@@ -735,7 +280,6 @@ public class PostRepositoryImpl implements PostRepository {
                 "ORDER BY GardeningPost.date DESC " +
                 "LIMIT ? OFFSET ?", gardeningPostRowMapper, pageable.getPageSize(), pageable.getOffset());
     }
-
 
     @Override
     public List<RecipePost> getAllRecipePost(Pageable pageable) {
